Convoluting_Images_SV- Description
This program takes in an input image and applies several filters to it. The filters include a small blur and large blur filter, a sharpening filter, and a laplacian filter, which detects the edges of certain objects. The user also gets to input their own values in a series of command-line arguments to create a personalized kernel and have it applied to the image.