Basic_Thresholding_SV- Description
This program takes in an image and a threshold value as inputs. First, the program runs the image through a simple binary and inverse binary thresholding, producing two images. Next, the program automatically generates a threshold value with otsu thresholding and applies the new threshold to the image, creating a total of four different thresholded images.