Imaging_Gradients_SV- Description
This program takes in an image as input. It converts it to grayscale before applying one of two functions to it. The user chooses which function is run. The first function derives a Sobel and Scharr convolution from the image and displays each created image. The second function displays the grayscale image as well as two plots of the gradient magnitude and orientation.