Edge_Detection_SV- Description
This program takes in an image as input as well as the parameters of the Canny edge detector. First, the image is converted to grayscale, and a bilateral filter is applied to it. Then, a final image is derived by applying the Canny edge detector, with the input parameters, to the image.