Connected_Components_SV- Description
This program takes in an input image and reveals the connected components of it. This program specializes in detecting letters from an image. The program first resizes the image to 745 pixels in the width by 960 pixels as the height as those dimensions are the easiest to work with. The program then converts the image to grayscale and binary and reduces high-frequency noise so that only the letters are detected. The program will print out the height, width, and area of all detected components and show images of every step in the process undertaken by this program to derive images detecting letters and components.